extract `MothershipController`; initialize onApplicationBootstrap
